requirements.txt
fastapi==0.115.5
uvicorn==0.34.2
alembic==1.14.0
python-dotenv==1.0.1
asyncpg==0.30.0
passlib==1.7.4
pydantic-settings==2.6.1
python-jose==3.4.0
pydantic[email]
SQLAlchemy==2.0.40